Flatten fake tests and restore sort order
