+++ docs/release-source/release/examples/fakes-01-using-fakes-instead-of-spies.test.js
@@ -3,7 +3,6 @@ const sinon = require("sinon");
 const referee = require("@sinonjs/referee");
 const assert = referee.assert;
 
-describe("FakeTest", function () {
     it("should be able to be used instead of spies", function () {
         const foo = {
             bar: () => "baz",
@@ -14,4 +13,3 @@ describe("FakeTest", function () {
         assert.equals(fake(), "baz"); // behaviour is the same
         assert.equals(fake.callCount, 1); // calling information is saved
     });
-});
