+++ docs/release-source/release/examples/fakes-02-using-fakes-instead-of-stubs.test.js
@@ -3,7 +3,6 @@ const sinon = require("sinon");
 const referee = require("@sinonjs/referee");
 const assert = referee.assert;
 
-describe("FakeTest", function () {
     it("should be able to be used instead of stubs", function () {
         const foo = {
             bar: () => "baz",
@@ -18,4 +17,3 @@ describe("FakeTest", function () {
         assert.equals(fake(), "fake value"); // returns fake value
         assert.equals(fake.callCount, 1); // saves calling information
     });
-});
