+++ docs/release-source/release/examples/fakes-03-creating-without-behaviour.test.js
@@ -3,7 +3,6 @@ const sinon = require("sinon");
 const referee = require("@sinonjs/referee");
 const assert = referee.assert;
 
-describe("FakeTest", function () {
     it("should create fake without behaviour", function () {
         // create a basic fake, with no behavior
         const fake = sinon.fake();
@@ -11,4 +10,3 @@ describe("FakeTest", function () {
         assert.isUndefined(fake()); // by default returns undefined
         assert.equals(fake.callCount, 1); // saves call information
     });
-});
