+++ docs/release-source/release/examples/fakes-04-creating-with-custom-behaviour.test.js
@@ -3,11 +3,9 @@ const sinon = require("sinon");
 const referee = require("@sinonjs/referee");
 const assert = referee.assert;
 
-describe("FakeTest", function () {
     it("should create fake with custom behaviour", function () {
         // create a fake that returns the text "foo"
         const fake = sinon.fake.returns("foo");
 
         assert.equals(fake(), "foo");
     });
-});
