+++ docs/release-source/release/examples/fakes-05-returns.test.js
@@ -3,10 +3,8 @@ const sinon = require("sinon");
 const referee = require("@sinonjs/referee");
 const assert = referee.assert;
 
-describe("FakeTest", function () {
     it("should create a fake that 'returns'", function () {
         const fake = sinon.fake.returns("apple pie");
 
         assert.equals(fake(), "apple pie");
     });
-});
