+++ docs/release-source/release/examples/fakes-10-firstArg.test.js
@@ -3,7 +3,6 @@ const sinon = require("sinon");
 const referee = require("@sinonjs/referee");
 const assert = referee.assert;
 
-describe("FakeTest", function () {
     it("should have working firstArg", function () {
         const f = sinon.fake();
         const date1 = new Date();
@@ -14,4 +13,3 @@ describe("FakeTest", function () {
 
         assert.isTrue(f.firstArg === date2);
     });
-});
